requirements.txt
Django==5.0
psycopg2==2.9.9
django-crispy-forms
django-leaflet
django-rest-framework
psycopg2-binary
